mcp_servers/requirements.txt
# MCP Server Requirements
mcp>=1.0.0
fastmcp>=0.1.0

# Django and database
django>=5.2.5
psycopg2-binary>=2.9.0
django-environ>=0.11.0
python-decouple>=3.8

# HTTP and API
httpx>=0.27.0
requests>=2.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Date/time handling
python-dateutil>=2.8.0

# Logging and monitoring
structlog>=23.0.0

# Async support
asyncio-mqtt>=0.16.0
aiofiles>=23.0.0

# Development tools
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0